examples/mobc: apply direcotory refactor migration(Drivers -> driver)
