+++ examples/mobc/src/src_user/applications/driver_instances/di_gs.c
@@ -8,7 +8,7 @@
 #include <src_core/tlm_cmd/common_cmd_packet_util.h>
 #include <src_core/library/print.h>
 #include <src_core/library/result.h>
-#include "../../Drivers/Com/gs_validate.h"
+#include "../../driver/com/gs_validate.h"
 #include "../../Settings/port_config.h"
 #include "../../Settings/DriverSuper/driver_buffer_define.h"
