+++ examples/mobc/src/src_user/applications/driver_instances/di_uart_test.h
@@ -5,7 +5,7 @@
 #ifndef DI_UART_TEST_H_
 #define DI_UART_TEST_H_
 
-#include "../../Drivers/Etc/uart_test.h"
+#include "../../driver/etc/uart_test.h"
 #include <src_core/system/application_manager/app_info.h>
 #include <src_core/tlm_cmd/common_cmd_packet.h>
