+++ examples/mobc/src/src_user/driver/aocs/aobc.h
@@ -6,7 +6,7 @@
 #define AOBC_H_
 
 #include <src_core/hal/uart.h>
-#include <src_core/Drivers/Super/driver_super.h>
+#include <src_core/driver/driver_super.h>
 #include <src_core/system/time_manager/obc_time.h>
 #include <src_core/tlm_cmd/common_cmd_packet.h>
 #include "./aobc_telemetry_data_definitions.h"
