+++ examples/mobc/src/src_user/driver/aocs/aobc_telemetry_definitions.h
