+++ examples/mobc/src/src_user/driver/com/gs_validate.c
