+++ examples/mobc/src/src_user/driver/etc/uart_test.c
