+++ examples/mobc/src/src_user/script/add_section/settings.pl
@@ -17,7 +17,7 @@ sub GetSetting {
 	$SETTING{'LOG_FILE'}  = './log.log';
 	$SETTING{'SEARCH_PATH'} = [
 		'applications',
-		'Drivers',
+		'driver',
 		'hal',
 		'library',
 		'Settings',
